Move JWT Auth mixin to `authentication` app
